make a separate Logger class
and move the logging interface into Main
