+++ README.md
@@ -139,7 +139,7 @@ libvips properties as properties of the PHP `Vips\Image` class.
 
 ```
 $ phpcs --standard=PSR2 src
-$ php ~/package/php/composer.phar install
+$ php ~/packages/php/composer.phar install
 $ vendor/bin/phpunit
 $ vendor/bin/phpdoc
 ```
